Update myArea.java
This program prompts a user for length and width to compute the area of a rectangle. It also tells the user that the shape is a square if the length and width are equal. There is a try catch exception to prompt the user to reenter an input if they did not enter an input, however this does not work and instead returns 0.