Create myarea2
This is a restructured version of my rectangle area calculator. The user prompt for a length or width is in class measurement. This program works if the user inputs equal measurements however if different measurements are used it is not return the user inputted integers and computes the wrong area.